requirements.txt
matplotlib==3.9.2
numpy==2.1.2
pandas==2.2.3
plotly==5.24.1
psutil==6.1.0
psycopg2-binary==2.9.10
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
toml==0.10.2
streamlit==1.39.0
load_dotenv